+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- PSR-14 event is dispatched when rate limit is exceeded (#1)
+
 ## [1.1.0] - 2023-04-01
 
 ### Added